project/requirements.txt
# <코드 1> requirements.txt
Flask==3.1.0
gradio==5.14.0
requests==2.32.3
chromadb==0.6.3
duckdb==1.1.3
openai==1.61.0
python-dotenv==1.0.1
# 필요 시 추가
# uvicorn==0.21.1  # FastAPI 쓰면 필요할 수 있음
# sentence-transformers==2.2.2  # 문서 임베딩
